+++ src/0_maze_gen/features.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/23 22:50:16 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/28 18:04:29 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/28 19:28:23 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,21 +14,21 @@
 
 void	remove_walls(char **maze, int height, int width)
 {
-	int	i;
-	int	j;
+	int	y;
+	int	x;
 
-	i = 2;
-	while (i < height - 4)
+	y = 1;
+	while (y < height - 2)
 	{
-		j = 2;
-		while (j < width - 4)
+		x = 1;
+		while (x < width - 2)
 		{
-			if (maze[i][j] == '1' && random_int(1, 100) < \
+			if (maze[y][x] == '1' && random_int(1, 100) < \
 								WALL_REMOVAL_CHANCE_PERCENT)
-				maze[i][j] = '0';
-			j++;
+				maze[y][x] = '0';
+			x++;
 		}
-		i++;
+		y++;
 	}
 }
 
@@ -57,15 +57,15 @@ static void	add_room(char **maze, t_scale start_pos, t_scale room_scale, t_scale
 	int	x;
 	int	y;
 
-	x = start_pos.x - 1;
-	while (++x < start_pos.x + room_scale.y)
+	y = start_pos.y - 1;
+	while (++y < start_pos.y + room_scale.y)
 	{
-		y = start_pos.y - 1;
-		while (++y < start_pos.y + room_scale.x)
+		x = start_pos.x - 1;
+		while (++x < start_pos.x + room_scale.x)
 		{
-			if (x > 0 && y > 0 && x < maze_size.y - 1 && \
-					y < maze_size.x - 1 && maze[x][y] == '1')
-				maze[x][y] = '0';
+			if (x > 0 && y > 0 && x < maze_size.x - 1 && \
+					y < maze_size.y - 1 && maze[y][x] == '1')
+				maze[y][x] = '0';
 		}
 	}
 }
@@ -75,12 +75,18 @@ void	add_rooms(char **maze, t_scale scale, int room_count)
 	t_scale	room_scale;
 	t_scale	start_pos;
 	int		i;
+	t_scale	max_room_size;
+	t_scale	min_room_size;
 
 	i = -1;
+	min_room_size.x = min(MIN_ROOM_SIZE, scale.x / MAX_ROOM_SIZE_REL_TO_MAP_SIZE);
+	min_room_size.y = min(MIN_ROOM_SIZE, scale.y / MAX_ROOM_SIZE_REL_TO_MAP_SIZE);
+	max_room_size.x = min(MAX_ROOM_SIZE, scale.x / MAX_ROOM_SIZE_REL_TO_MAP_SIZE);
+	max_room_size.y = min(MAX_ROOM_SIZE, scale.y / MAX_ROOM_SIZE_REL_TO_MAP_SIZE);
 	while (++i < room_count)
 	{
-		room_scale.x = random_int(MIN_ROOM_SIZE, MAX_ROOM_SIZE);
-		room_scale.y = random_int(MIN_ROOM_SIZE, MAX_ROOM_SIZE);
+		room_scale.x = random_int(min_room_size.x, max_room_size.x);
+		room_scale.y = random_int(min_room_size.y, max_room_size.y);
 		start_pos.x = random_int(1, scale.x - room_scale.y - 2);
 		start_pos.y = random_int(1, scale.y - room_scale.x - 2);
 		if (room_scale.x <= 0 || room_scale.y <= 0)