+++ src/0_maze_gen/mapmaker.c
@@ -6,13 +6,13 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/23 22:46:53 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/28 18:13:42 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/28 19:29:26 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/cub3d.h"
 
-static void	get_shuffled_directions(int dx[], int dy[])
+static inline void	get_shuffled_directions(int dx[], int dy[])
 {
 	int	i;
 	int	r;
@@ -47,41 +47,45 @@ static void	generate_maze(char **maze, bool **visited, t_scale pos, t_scale size
 	int		dy[4];
 	int		i;
 	t_scale	new_pos;
+	t_scale	mid_pos;
 
 	get_shuffled_directions(dx, dy);
-	visited[pos.x][pos.y] = true;
-	maze[pos.x][pos.y] = '0';
+	visited[pos.y][pos.x] = true;
+	maze[pos.y][pos.x] = '0';
 	i = -1;
 	while (++i < 4)
 	{
 		new_pos.x = pos.x + dx[i];
 		new_pos.y = pos.y + dy[i];
-		if (new_pos.x > 0 && new_pos.x < size.y - 1 && new_pos.y > 0 && \
-					new_pos.y < size.x - 1 && !visited[new_pos.x][new_pos.y])
+		mid_pos.x = (pos.x + new_pos.x) / 2;
+		mid_pos.y = (pos.y + new_pos.y) / 2;
+		if (new_pos.x > 0 && new_pos.x < size.x - 1 && new_pos.y > 0 && \
+					new_pos.y < size.y - 1 && !visited[new_pos.y][new_pos.x] && \
+					mid_pos.x > 0 && mid_pos.x < size.x - 1 && mid_pos.y > 0 && mid_pos.y < size.y - 1)
 		{
-			maze[(pos.x + new_pos.x) / 2][(pos.y + new_pos.y) / 2] = '0';
+			maze[mid_pos.y][mid_pos.x] = '0';
 			generate_maze(maze, visited, new_pos, size);
 		}
 	}
 }
 
-static void	setup_maze(char ***maze, bool ***visited, t_scale map_scale)
+static inline void	setup_maze(char ***maze, bool ***visited, t_scale map_scale)
 {
-	int	i;
-	int	j;
+	int	x;
+	int	y;
 
 	*maze = gc_malloc(map_scale.y * sizeof(char *));
 	*visited = gc_malloc(map_scale.y * sizeof(bool *));
-	i = -1;
-	while (++i < map_scale.y)
+	y = -1;
+	while (++y < map_scale.y)
 	{
-		j = -1;
-		(*maze)[i] = gc_malloc(map_scale.x * sizeof(char));
-		(*visited)[i] = gc_malloc(map_scale.x * sizeof(bool));
-		while (++j < map_scale.x)
+		(*maze)[y] = gc_malloc(map_scale.x * sizeof(char));
+		(*visited)[y] = gc_malloc(map_scale.x * sizeof(bool));
+		x = -1;
+		while (++x < map_scale.x)
 		{
-			(*maze)[i][j] = '1';
-			(*visited)[i][j] = false;
+			(*maze)[y][x] = '1';
+			(*visited)[y][x] = false;
 		}
 	}
 }
@@ -94,15 +98,14 @@ void	generate_map(void)
 
 	map_scale = (t_scale){random_int(MIN_MAP_SIZE, MAX_MAP_SIZE), random_int(MIN_MAP_SIZE, MAX_MAP_SIZE)};
 	setup_maze(&maze, &visited, map_scale);
-	write(STDOUT_FILENO, "Generating maze...\n", 19);
-	generate_maze(maze, visited, (t_scale){random_int(1, map_scale.x - 2), random_int(1, map_scale.x - 2)}, map_scale);
-	write(STDOUT_FILENO, "Maze generated\n", 15);
+	generate_maze(maze, visited, (t_scale){random_int(1, map_scale.y - 2), random_int(1, map_scale.x - 2)}, map_scale);
 	place_player_spawn(maze, map_scale);
-	write(STDOUT_FILENO, "Player spawn placed\n", 21);
+	logger(LOGGER_STEP, "Player spawn placed");
 	remove_walls(maze, map_scale.y, map_scale.x);
-	write(STDOUT_FILENO, "Walls removed\n", 14);
+	logger(LOGGER_STEP, "Walls removed");
 	add_rooms(maze, map_scale, random_int(MIN_ROOMS_AMOUNT, MAX_ROOMS_AMOUNT));
-	write(STDOUT_FILENO, "Rooms added\n", 12);
+	logger(LOGGER_STEP, "Rooms added");
 	write_cub_file(maze, map_scale.y, map_scale.x, "generated_map.cub");
-	write(STDOUT_FILENO, "Map saved to generated_map.cub\n", 32);
+	logger(LOGGER_INFO, "Map saved to generated_map.cub");
+	gc_cleanup_and_reinit();
 }